+++ test/lit.cfg
@@ -274,7 +274,7 @@ config.inferSwiftBinary = inferSwiftBinary
 config.swift = inferSwiftBinary('swift')
 config.swiftc = inferSwiftBinary('swiftc')
 config.sil_opt = inferSwiftBinary('sil-opt')
-config.sil_extract = inferSwiftBinary('sil-extract')
+config.sil_func_extractor = inferSwiftBinary('sil-func-extractor')
 config.sil_nm = inferSwiftBinary('sil-nm')
 config.lldb_moduleimport_test = inferSwiftBinary('lldb-moduleimport-test')
 config.swift_ide_test = inferSwiftBinary('swift-ide-test')
@@ -356,7 +356,7 @@ config.substitutions.append( ('%swiftc_driver_plain', "%r" % config.swiftc) )
 config.substitutions.append( ('%swift_driver', "env SDKROOT= %r %s %s" % (config.swift, mcp_opt, config.swift_test_options)) )
 config.substitutions.append( ('%swiftc_driver', "env SDKROOT= %r %s %s" % (config.swiftc, mcp_opt, config.swift_test_options)) )
 config.substitutions.append( ('%sil-opt', "%r %s" % (config.sil_opt, mcp_opt)) )
-config.substitutions.append( ('%sil-extract', "%r %s" % (config.sil_extract, mcp_opt)) )
+config.substitutions.append( ('%sil-func-extractor', "%r %s" % (config.sil_func_extractor, mcp_opt)) )
 config.substitutions.append( ('%sil-nm', "%r %s" % (config.sil_nm, mcp_opt)) )
 config.substitutions.append( ('%lldb-moduleimport-test', "%r %s" % (config.lldb_moduleimport_test, mcp_opt)) )
 config.substitutions.append( ('%swift-ide-test_plain', config.swift_ide_test) )
@@ -428,7 +428,7 @@ disallow('swiftc')
 disallow('swift_driver')
 disallow('swiftc_driver')
 disallow('sil-opt')
-disallow('sil-extract')
+disallow('sil-func-extractor')
 disallow('sil-nm')
 disallow('lldb-moduleimport-test')
 disallow('swift-ide-test')
@@ -937,9 +937,9 @@ config.target_parse_verify_swift = (
     '%s -typecheck -verify -disable-objc-attr-requires-foundation-module %%s'
     % (config.target_swift_frontend, ))
 
-config.target_sil_extract = (
+config.target_sil_func_extractor = (
     '%s -target %s %s'
-    % (config.sil_extract, config.variant_triple, mcp_opt))
+    % (config.sil_func_extractor, config.variant_triple, mcp_opt))
 
 config.target_sil_nm= (
     '%s -target %s %s'
@@ -994,7 +994,7 @@ config.substitutions.append(('%scale-test',
                              '{} --swiftc-binary={} --tmpdir=%t'.format(
                                  config.scale_test, config.swiftc)))
 config.substitutions.append(('%target-sil-opt', config.target_sil_opt))
-config.substitutions.append(('%target-sil-extract', config.target_sil_extract))
+config.substitutions.append(('%target-sil-func-extractor', config.target_sil_func_extractor))
 config.substitutions.append(('%target-sil-nm', config.target_sil_nm))
 config.substitutions.append(
     ('%target-swift-ide-test\(mock-sdk:([^)]+)\)',
